backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
resend>=2.0.0
apscheduler>=3.10.0
PyJWT[crypto]>=2.8.0
